Move the TOC to after the introduction
This moves the TOC to after the introductory paragraphs, so that someone who
lands on the repo gets the summary paragraphs first (to find out what the
project is), before getting the overview of the document structure.

Also, I removed the TOC section for the rough notes, because highlighting them
at the top is less important (they should all be merged into the document
eventually anyways, rather than existing as placeholder artefacts).